DescriptionDescriptor: interface and Default implementation
